==> comparison_Yazar_2022/step4_scatterplots_Yazar.Rmd <==

---
title: "Scatterplots of -log10(pvalue) for the eGenes from Yazar 2022"
author: "Si Liu and Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

For the eGenes from Yazar et al. (2022), we compare -log10(p-value) from Yazar et al. versus  -log10(p-value) by CSeQTL from GTEx whole blood bulk tissues. We first identify  the genes that are considered in the eQTL mapping by both Yazar et al. and CSeQTL, and then extract p-values for those Yazar eGenes. 


# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)

theme_set(theme_classic())

```

## Data directory and file names
```{r}
CSeQTL_dir = "../../CSeQTL/data2share/"
GTEx_eqtl_file = "GTExBlood_eqtls.rds"

reference_dir = "../../CSeQTL/Yazar_2022/"

Yazar_2022_kept_folder = "Yazar_eSNP1_50kb/"
Yazar_2022_considered  = "Yazar_gene_considered.rds"
```

## cell type names
```{r}
Yazar_cell_types = c("B IN", "B Mem", "CD4 NC", "CD4 ET", "CD4 SOX4", 
                     "CD8 NC", "CD8 ET", "CD8 S100B", "DC", "Mono C", 
                     "Mono NC", "NK R", "NK", "Plasma")

CSeQTL_cell_types = c("B", "CD4T", "CD8T", "Monocyte", "NK")
```

# Get overlap

## Step 1: load the genes considered for each cell type by Yazar et al. (2022). 

```{r}
gene_con_Yazar_list = readRDS(paste0(reference_dir, Yazar_2022_considered))
length(gene_con_Yazar_list)

```

## Step 2: load the genes considered under each cell type by CSeQTL.

```{r}
GTExBlood <- readRDS(paste0(CSeQTL_dir, GTEx_eqtl_file))
length(GTExBlood)
names(GTExBlood)

lapply(GTExBlood, dim)

GTExBlood = GTExBlood[["eqtl"]]
dim(GTExBlood)
head(GTExBlood)
round(colSums(is.na(GTExBlood))/nrow(GTExBlood), 3)

length(unique(GTExBlood$ENSG))
length(unique(GTExBlood$GENE))
length(unique(GTExBlood$SNP))

table(GTExBlood$CELLTYPE, useNA = "ifany")

GTExBlood$ENSG_version = GTExBlood$ENSG
GTExBlood$ENSG = gsub("\\..*", "", GTExBlood$ENSG_version)

GTExBlood = GTExBlood[which((GTExBlood$trim==TRUE) & 
                              (GTExBlood$perm==FALSE) & 
                              (GTExBlood$MODEL=="cseqtl") & 
                              (!is.na(GTExBlood$PVAL))), ]

dim(GTExBlood)

n_con_CSeQTL_list = list()
gene_con_CSeQTL_list = list()

for (j in 1:length(CSeQTL_cell_types)){
  
  cct = CSeQTL_cell_types[j]
  
  cct_GTExBlood = GTExBlood[which(GTExBlood$CELLTYPE == cct), ]
  cct_considered_CSeQTL = unique(cct_GTExBlood$ENSG)
  
  n_con_CSeQTL_list[[cct]] = length(cct_considered_CSeQTL)
  gene_con_CSeQTL_list[[cct]] = cct_considered_CSeQTL
  
}

unlist(n_con_CSeQTL_list)
```


## Step 3: generate scatter plot for each pair of cell types

```{r}

n_Yazar_eGene_kept_mat =  matrix(NA, ncol=length(CSeQTL_cell_types), 
                                 nrow=length(Yazar_cell_types))

pdf(width = 12, height = 6, "figures/step4_pvalue_scatterplots.pdf")

for (i in 1:length(Yazar_cell_types)){
  
  Yazar_ct = Yazar_cell_types[i]
  Yazar_ct_string = sub(" ", "_", Yazar_ct)
  
  Yazar_ct_eGenes_path = paste0(reference_dir, Yazar_2022_kept_folder, 
                                "Yazar_eSNP1_50kb_", Yazar_ct_string, ".csv")
  
  df_Yazar_ct_eGenes = read.csv(Yazar_ct_eGenes_path, header = TRUE)
  
  Yazar_ct_eGenes = df_Yazar_ct_eGenes$Gene.Ensembl.ID
  
  p_ct_list = list()
  
  for (j in 1:length(CSeQTL_cell_types)){
    
    CSeQTL_ct = CSeQTL_cell_types[j]
    
    Yazar_ct_considered  = gene_con_Yazar_list[[Yazar_ct_string]]
    CSeQTL_ct_considered = gene_con_CSeQTL_list[[CSeQTL_ct]]
    
    overlap_considered = intersect(Yazar_ct_considered, CSeQTL_ct_considered)
    
    Yazar_ct_eGenes_kept = intersect(Yazar_ct_eGenes, overlap_considered)
    
    n_Yazar_eGene_kept_mat[i, j] = length(Yazar_ct_eGenes_kept)
    
    Yazar_eGenes_Yazar_pval = 
      df_Yazar_ct_eGenes[match(Yazar_ct_eGenes_kept, 
                               df_Yazar_ct_eGenes$Gene.Ensembl.ID), ]$pvalue
    
    CSeQTL_ct_GTExBlood = GTExBlood[which(GTExBlood$CELLTYPE == CSeQTL_ct), ]
    
    Yazar_eGenes_CSeQTL_pval = 
      CSeQTL_ct_GTExBlood[match(Yazar_ct_eGenes_kept, 
                                CSeQTL_ct_GTExBlood$ENSG), ]$PVAL
    
    check = length(Yazar_ct_eGenes_kept) == 
      sum(!is.na(Yazar_eGenes_CSeQTL_pval))
    stopifnot(check)

    df_ct_plot_pvalues = data.frame(Yazar = Yazar_eGenes_Yazar_pval, 
                                    CSeQTL = Yazar_eGenes_CSeQTL_pval)
    
    cur_corr = 
      cor(-log10(df_ct_plot_pvalues$CSeQTL), 
          -log10(df_ct_plot_pvalues$Yazar), method = "spearman")
    
    p_ct_list[[j]] = ggplot(df_ct_plot_pvalues, 
                            aes(x = -log10(CSeQTL), y = -log10(Yazar))) + 
      geom_point(size = 1, alpha = 0.5) + 
      geom_abline(intercept = 0, slope = 1) + 
      xlab("-log10(CSeQTL)") + ylab("-log10(Yazar)") + 
      ggtitle(paste0("Yazar: ", Yazar_ct_string, "\nCSeQTL: ", CSeQTL_ct, 
                     "\nSpearman corr: ", signif(cur_corr,2)))
  }
  
  print(ggarrange(plotlist=p_ct_list, nrow = 2, ncol = 4))
  
}

dev.off()

```


## Print out sanity check matrices

```{r}
rownames(n_Yazar_eGene_kept_mat) = Yazar_cell_types
colnames(n_Yazar_eGene_kept_mat) = CSeQTL_cell_types

n_Yazar_eGene_kept_mat

```

# Session information
```{r}
gc()

sessionInfo()
```
