==> comparison_Yazar_2022/step6_compare_eGenes.Rmd <==
---
title: "Check the difference between sc-eQTL results and CSeQTL"
author: "Si Liu and Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

Check the difference between the eGenes from Yazar et al. (2022) and eGenes identified by CSeQTL from GTEx whole blood RNA-seq data.

# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)
theme_set(theme_classic())

params = list(cs_q_cutoff=5e-3, fold_cutoff=1)
cs_q_cutoff = params$cs_q_cutoff
fold_cutoff = params$fold_cutoff

cat("cs_q_cutoff: ",  cs_q_cutoff,  "\n")
cat("fold_cutoff: ",  fold_cutoff,  "\n")

config = sprintf("cs_q_%.0e_fold_%.1f", 
                 cs_q_cutoff, fold_cutoff)
config
```

## Data directory and file names

```{r}
CSeQTL_dir     = "../../CSeQTL/data2share/"
GTEx_eqtl_file = "GTExBlood_eqtls.rds"

Yazar_2022_dir = "../../CSeQTL/Yazar_2022/"

Yazar_2022_kept_folder = "Yazar_eSNP1_50kb/"
Yazar_2022_genes       = "Yazar_gene_considered.rds"
```

# Read data/results and make comparison

## Read eGenes for comparison. 

```{r}
fnm = sprintf("results/step3_%s_eGenes.rds", config)
eGenes_overlap_list = readRDS(fnm)
sapply(eGenes_overlap_list, function(x){
  sapply(x, function(z){sapply(z, length)})})
```

## Cell type names

```{r}
cs_cell_types = c("B", "CD4T", "CD8T", "Monocyte", "NK")

ot_cell_types = c("B IN", "B Mem", "CD4 NC", "CD4 ET", 
                     "CD8 NC", "CD8 ET", "CD8 S100B", "Mono C", 
                     "Mono NC", "NK R", "NK")

ct_groups = list(B = c("B IN", "B Mem"), 
                 CD4T = c("CD4 NC", "CD4 ET"),
                 CD8T = c("CD8 NC", "CD8 ET", "CD8 S100B"),
                 Monocyte = c("Mono C", "Mono NC"),
                 NK = c("NK R", "NK"))

```


## Read in Yazar 2022 results

```{r fig.width=9, fig.height=12, warning = FALSE}

plist = list()

for (i in 1:length(ot_cell_types)){
  
  ot_ct = ot_cell_types[i]
  ot_ct_str = sub(" ", "_", str_squish(sub("/", "", ot_ct)))
  
  Yazar_ct_eGenes_path = paste0(Yazar_2022_dir, Yazar_2022_kept_folder, 
                            "Yazar_eSNP1_50kb_", ot_ct_str, ".csv")
  
  df_Yazar_ct  = read.csv(Yazar_ct_eGenes_path, header = TRUE)
  
  mat_i = match(eGenes_overlap_list[[ot_ct]][[1]]$Yazar, 
                df_Yazar_ct$Gene.Ensembl.ID)
  stopifnot(sum(is.na(mat_i)) == 0)
  
  df_Yazar_ct = df_Yazar_ct[mat_i,]
  df_Yazar_ct$CSeQTL = 
    df_Yazar_ct$Gene.Ensembl.ID %in% eGenes_overlap_list[[ot_ct]][[1]]$CSeQTL
  table(df_Yazar_ct$CSeQTL)
  
  ks1 = ks.test(log10(df_Yazar_ct$dist[!df_Yazar_ct$CSeQTL] + 1), 
                log10(df_Yazar_ct$dist[df_Yazar_ct$CSeQTL]+1))
  
  plist[[3*(i-1)+1]] = ggplot(df_Yazar_ct, 
                              aes(x=CSeQTL, y=log10(dist+1), color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(sprintf("%s, %d vs. %d", ot_ct, sum(!df_Yazar_ct$CSeQTL), 
                    sum(df_Yazar_ct$CSeQTL)), 
            subtitle=sprintf("KS test p-val=%.1e", ks1$p.value))
  
  df_Yazar_ct$qvalue[which(df_Yazar_ct$qvalue < 1e-50)] = 1e-50

  ks1 = ks.test(-log10(df_Yazar_ct$qvalue)[!df_Yazar_ct$CSeQTL], 
                -log10(df_Yazar_ct$qvalue)[df_Yazar_ct$CSeQTL])

  
  plist[[3*(i-1)+2]] = ggplot(df_Yazar_ct, aes(x=CSeQTL, y=-log10(qvalue), 
                                               color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

  ks1 = ks.test(-log10(df_Yazar_ct$FDR)[!df_Yazar_ct$CSeQTL], 
                -log10(df_Yazar_ct$FDR)[df_Yazar_ct$CSeQTL])

  plist[[3*(i-1)+3]] = ggplot(df_Yazar_ct, aes(x=CSeQTL, 
                                               y=-log10(FDR), 
                                               color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

}

length(plist)

ggarrange(plotlist=plist[1:12], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[13:21], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[22:33], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")
```

## Read in CSeQTL results

```{r fig.width=9, fig.height=12, warning = FALSE}

GTExWB <- readRDS(paste0(CSeQTL_dir, GTEx_eqtl_file))
length(GTExWB)
names(GTExWB)

lapply(GTExWB, dim)

GTExWB = GTExWB[["eqtl"]]
dim(GTExWB)
GTExWB[1:2,]

table(GTExWB$CELLTYPE, useNA = "ifany")

GTExWB$ENSG_version = GTExWB$ENSG
GTExWB$ENSG = gsub("\\..*","", GTExWB$ENSG_version)

GTExWB = GTExWB[which((GTExWB$trim==TRUE) & 
                              (GTExWB$perm==FALSE) & 
                              (GTExWB$MODEL=="cseqtl") & 
                              (!is.na(GTExWB$PVAL))), ]

plist = list()

k = 1
for (i in 1:length(cs_cell_types)){
  
  cct = cs_cell_types[i]
  df_GTEx_i = GTExWB[which(GTExWB$CELLTYPE == cct), ]
  
  for (ct_ot in ct_groups[[cct]]){
    mat_i = match(eGenes_overlap_list[[ct_ot]][[1]]$CSeQTL, df_GTEx_i$ENSG)
    stopifnot(sum(is.na(mat_i)) == 0)
    
    df_GTEx = df_GTEx_i[mat_i,]
    df_GTEx$sc_eGene = df_GTEx$ENSG %in% eGenes_overlap_list[[ct_ot]][[1]]$Yazar
    table(df_GTEx$sc_eGene)
    
    ks1 = ks.test(log10(df_GTEx$MUa[!df_GTEx$sc_eGene]), 
                  log10(df_GTEx$MUa[df_GTEx$sc_eGene]))
    
    plist[[k]] = ggplot(df_GTEx, 
                        aes(x=sc_eGene, y=log10(MUa), color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
      ggtitle(sprintf("%s, %d vs. %d", ct_ot, 
                      sum(!df_GTEx$sc_eGene), 
                      sum(df_GTEx$sc_eGene)), 
              subtitle=sprintf("KS test p-val=%.1e", ks1$p.value))
    k = k + 1
    
    df_GTEx$qvalue[which(df_GTEx$qvalue < 1e-50)] = 1e-50
    
    ks1 = ks.test(-log10(df_GTEx$qvalue)[!df_GTEx$sc_eGene], 
                  -log10(df_GTEx$qvalue)[df_GTEx$sc_eGene])
  
    plist[[k]] = ggplot(df_GTEx, aes(x=sc_eGene, 
                                     y=-log10(qvalue), 
                                     color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
      ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))
    k = k + 1

    ks1 = ks.test(abs(log10(df_GTEx$ETA))[!df_GTEx$sc_eGene], 
                  abs(log10(df_GTEx$ETA))[df_GTEx$sc_eGene])
  
    plist[[k]] = ggplot(df_GTEx, aes(x=sc_eGene, 
                                     y=abs(log10(ETA)), 
                                     color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
      ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))
    k = k + 1
  }
}

length(plist)

ggarrange(plotlist=plist[1:12], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[13:21], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[22:33], nrow = 4, ncol = 3, 
          common.legend=TRUE, legend="top")
```

# Session information
```{r}
gc()

sessionInfo()
```
