==> comparison_Yazar_2022/step8_summerize_eQTL_directions.Rmd <==
---
title: "Check whethre eQTL effects have the same directions"
author: "Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

Check the difference between the eGenes from Yazar et al. (2022) and eGenes identified by CSeQTL from GTEx whole blood RNA-seq data.

# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)
theme_set(theme_classic())
```

# Check effect directions

## Read in overlap and summarize results

```{r, fig.width=5, fig.height=4}
overlap = fread("results/step7b_overlap.tsv")
dim(overlap)
overlap[1:2,]

table(is.na(overlap$lm_beta))
overlap = overlap[which(!is.na(overlap$lm_beta)),]
dim(overlap)
overlap[1:2,]
table(is.na(overlap$spearman))

g1 = ggplot(overlap, aes(x=spearman, y=lm_beta, color=cell_type_ot)) + 
         geom_point() + 
  geom_hline(yintercept=0, linetype="dashed") + 
  geom_vline(xintercept=0, linetype="dashed")
g1

ggplot(overlap, aes(x=-log10(spearman_pval), y=-log10(lm_pval), 
                    color=cell_type_ot)) + 
         geom_point() + 
  geom_hline(yintercept=0, linetype="dashed") + 
  geom_vline(xintercept=0, linetype="dashed")


table(overlap$spearman > 0, overlap$lm_beta > 0)
table(overlap$spearman > 0, overlap$ETA > 1)


c1 = chisq.test(overlap$lm_beta > 0, overlap$ETA > 1)
c1$p.value
c1$expected
c1$observed

sum(diag(c1$expected)/nrow(overlap))
sum(diag(c1$observed)/nrow(overlap))

cutoffs = data.frame(qvalue = rep(c(0.10, 0.005), each=2), 
                     lm_pval = rep(c(0.05, 0.01), times=2))
cutoffs

v1 = rep(NA, nrow(cutoffs))
res0 = data.frame(cutoffs, neQTL = v1, expected=v1, observed=v1, 
                  chisq_pval=v1)

for(i in 1:nrow(cutoffs)){
  qi = cutoffs$qvalue[i]
  pi = cutoffs$lm_pval[i]
  
  w2kp = which(overlap$qvalue < qi & overlap$lm_pval < pi)

  c1 = chisq.test(overlap$lm_beta[w2kp] > 0, overlap$ETA[w2kp] > 1)
  c1$expected
  c1$observed

  res0$neQTL[i]    = length(w2kp)
  res0$expected[i] = round(sum(diag(c1$expected))/length(w2kp),2)
  res0$observed[i] = round(sum(diag(c1$observed))/length(w2kp),2)
  res0$chisq_pval[i] = signif(c1$p.value,2)
}

res0

```

## Summarize overlap for each CSeQTL cell type

```{r}
signal1 = overlap$qvalue < 0.1 & overlap$lm_pval < 0.01
table(overlap$CELLTYPE, signal1)

signal2 = overlap$qvalue < 0.005 & overlap$lm_pval < 0.01
table(overlap$CELLTYPE, signal2)

cts = unique(overlap$CELLTYPE)
length(cts)

v1 = rep(NA, 2*length(cts))
res = data.frame(cell_type = rep(cts, each=2), 
                 qvalue = rep(c(0.1, 0.005), times=length(cts)),
                 neQTL = v1, expected=v1, observed=v1, chisq_pval=v1)

for(i in 1:nrow(res)){
  ct1  = res$cell_type[i]
  qval = res$qvalue[i]
  
  w2kp = overlap$qvalue < qval & overlap$lm_pval < 0.01
  w2kp = which(w2kp & overlap$CELLTYPE == ct1)
  
  c1 = chisq.test(overlap$lm_beta[w2kp] > 0, overlap$ETA[w2kp] > 1)
  
  res$neQTL[i]    = length(w2kp)
  res$expected[i] = round(sum(diag(c1$expected))/length(w2kp),2)
  res$observed[i] = round(sum(diag(c1$observed))/length(w2kp),2)
  res$chisq_pval[i] = signif(c1$p.value,2)
}

res
```


## Summarize overlap for each Yazar et al. cell type

```{r}
signal1 = overlap$qvalue < 0.1 & overlap$lm_pval < 0.01
table(overlap$cell_type_ot, signal1)

signal2 = overlap$qvalue < 0.005 & overlap$lm_pval < 0.01
table(overlap$cell_type_ot, signal2)

cts = unique(overlap$cell_type_ot)
length(cts)

v1 = rep(NA, 2*length(cts))
res2 = data.frame(cell_type = rep(cts, each=2), 
                 qvalue = rep(c(0.1, 0.005), times=length(cts)),
                 neQTL = v1, expected=v1, observed=v1, chisq_pval=v1)

for(i in 1:nrow(res2)){
  ct1  = res2$cell_type[i]
  qval = res2$qvalue[i]
  
  w2kp = overlap$qvalue < qval & overlap$lm_pval < 0.01
  w2kp = which(w2kp & overlap$cell_type_ot == ct1)
  
  c1 = suppressWarnings(chisq.test(overlap$lm_beta[w2kp] > 0, 
                                   overlap$ETA[w2kp] > 1))
  
  res2$neQTL[i]    = length(w2kp)
  res2$expected[i] = round(sum(diag(c1$expected))/length(w2kp),2)
  res2$observed[i] = round(sum(diag(c1$observed))/length(w2kp),2)
  res2$chisq_pval[i] = signif(c1$p.value,2)
}

res2
```

## plot
```{r fig.width=4.5, fig.height=3}
theme1 = theme(axis.text.x = element_text(angle = 60, vjust = 1, 
                                   size = 11, hjust = 1),
               axis.text.y = element_text(size = 11),
               legend.text=element_text(size=11), 
               axis.title.x = element_blank())

res$qvalue = as.factor(res$qvalue)
p <- ggplot(res, aes(x=cell_type, y=observed, fill=qvalue)) +
  geom_bar(stat="identity", position=position_dodge()) + 
  ylab("prop. consistent eQTL direction") 
p + scale_fill_brewer(palette="Dark2") + theme1
```

## Save results
```{r}
res = rbind(res, res2)
fwrite(res, file="results/step8_eQTL_directions_by_cell_type.tsv", sep="\t")

fwrite(res0, file="results/step8_eQTL_directions.tsv", sep="\t")

```



# Session information
```{r}
gc()

sessionInfo()
```
